Update Styles and Readme
