+++ src/styles/BookList.module.css
@@ -1,7 +1,3 @@
-#root {
-  background-color: #fafafa;
-}
-
 /* For BookList.jsx */
 .booklist {
   background-color: #fafafa;
@@ -17,9 +13,9 @@
   justify-content: space-between;
   border-radius: 4px;
   border: 1px solid #e8e8e8;
-  margin: 2rem;
+  margin: 2%;
   background-color: #fff;
-  padding: 1rem;
+  padding: 1%;
 }
 
 .bookInnerContainer {
@@ -67,13 +63,15 @@
 }
 
 .bookTitle {
-  font-family: Montserrat;
-  font-size: 1.5rem;
+  height: 1.813rem;
+  font-family: RobotoSlab;
+  font-size: 1.375rem;
   font-weight: bold;
   font-stretch: normal;
   font-style: normal;
   line-height: normal;
-  letter-spacing: normal;
+  letter-spacing: -0.2px;
+  color: #121212;
 }
 
 .bookAuthor {
@@ -119,7 +117,6 @@
 .oval {
   width: 4.25rem;
   height: 4.25rem;
-  border-style: solid;
   border-width: 5px;
   border-image-source: linear-gradient(to bottom, var(--medium-blue), #379cf6);
   border-image-slice: 1;
@@ -155,8 +152,14 @@
   color: black;
 }
 
+.currentContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin: 4.9% 0;
+}
 
-currentChapter {
+.currentChapter {
   width: 7.875rem;
   height: 1.125rem;
   opacity: 0.5;
@@ -188,14 +191,5 @@ currentChapter {
   height: 2.063rem;
   border-radius: 3px;
   background-color: #0290ff;
-  width: 8.938rem;
-  font-family: RobotoSlab;
-  font-size: 0.813rem;
-  font-weight: 300;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: normal;
-  letter-spacing: 0.5px;
-  text-align: center;
-  color: var(--white-two);
+  color: white;
 }
