+++ src/styles/CreateBook.module.css
@@ -1,0 +1,55 @@
+.createContainer {
+  margin-left: 10%;
+  height: 170px;
+}
+
+.createBookForm {
+  display: flex;
+  flex-direction: row;
+  gap: 50px;
+  width: 100%;
+  padding: 20px;
+  background-color: #f5f5f5;
+}
+
+.createBookForm label {
+  font-size: 1.2em;
+  color: #888;
+}
+
+.createBookForm input, .createBookForm select {
+  width: 100%;
+  padding: 10px;
+  font-size: 1em;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  background-color: #fff;
+}
+
+.createBookForm button {
+  padding: 10px 20px;
+  font-size: 1.2em;
+  color: #fff;
+  background-color: #0290ff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.createBookForm button:hover {
+  background-color: #0270a1;
+}
+
+.addBookTitle {
+  width: 10.875rem;
+  height: 1.5rem;
+  margin: 1.813rem 32.75rem 1.188rem 0;
+  font-family: Montserrat;
+  font-size: 1.25rem;
+  font-weight: bold;
+  font-stretch: normal;
+  font-style: normal;
+  line-height: normal;
+  letter-spacing: -0.18px;
+  color: #888;
+}
