+++ src/styles/NavBar.module.css
@@ -4,6 +4,7 @@
   justify-content: space-between;
   padding-left: 9%;
   height: 5.938rem;
+  background-color: white;
 }
 
 .container {
@@ -44,9 +45,9 @@
 .navLink {
   list-style: none;
   text-decoration: none;
-  width: 3.688rem;
+  /* width: 3.688rem; */
   height: 1rem;
-  margin: 1.125rem 2.563rem 0.688rem 0;
+  /* margin: 1.125rem 2.563rem 0.688rem 0; */
   font-family: Montserrat;
   font-size: 0.813rem;
   font-weight: normal;
